// lib/api/schema.d.ts
export interface paths {
// ...
	"/latest.json": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/** Get the latest topics */
		get: operations["listLatestTopics"];
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
}
export type webhooks = Record<string, never>;
export interface components {
	schemas: never;
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
	getBackups: {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			/** @description success response */
			200: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					"application/json": {
						filename: string;
						size: number;
						last_modified: string;
					}[];
				};
			};
		};
	};
// ...
	listCategoryTopics: {
		parameters: {
			query?: never;
			header?: never;
			path: {
				slug: string;
				id: number;
			};
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			/** @description success response */
			200: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					"application/json": {
						users?: {
							id: number;
							username: string;
							name: string;
							avatar_template: string;
						}[];
						primary_groups?: unknown[];
						topic_list: {
							can_create_topic: boolean;
							per_page: number;
							top_tags?: unknown[];
							topics: {
								id: number;
								title: string;
								fancy_title: string;
								slug: string;
								posts_count: number;
								reply_count: number;
								highest_post_number: number;
								image_url: string | null;
								created_at: string;
								last_posted_at: string;
								bumped: boolean;
								bumped_at: string;
								archetype: string;
								unseen: boolean;
								pinned: boolean;
								unpinned: string | null;
								excerpt: string;
								visible: boolean;
								closed: boolean;
								archived: boolean;
								bookmarked: string | null;
								liked: string | null;
								views: number;
								like_count: number;
								has_summary: boolean;
								last_poster_username: string;
								category_id: number;
								pinned_globally: boolean;
								featured_link: string | null;
								posters: {
									extras: string;
									description: string;
									user_id: number;
									primary_group_id: number | null;
								}[];
							}[];
						};
					};
				};
			};
		};
	};
// ...
}